+++ src/classes/enemy-dgmn.js
@@ -1,8 +1,0 @@
-class EnemyDgmn extends Dgmn{
-  constructor(...args){
-    super(...args);
-  }
-
-  // STUFF SPECIFIC TO ENEMIES
-  
-}