+++ src/data/attacks.db.js
@@ -1,0 +1,7 @@
+export const attacksDB = {
+  babyFlame: {
+    displayName: 'Baby Flame',
+    type: 'fire',
+    targets: 'single'
+  }
+}