+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 # DIGIMON : TOWER OF RUIN
 
-## CURRENT VERSION : 0.0.0-alpha.2
+## CURRENT VERSION : 0.0.0-alpha.5
 
 ### 0.0.0
 Setting up all of the basic structure for the system, until the game is in any way testable with other users
+
+#### v0.0.0-alpha.5
+* Built Battle Menu
+  * Create Status Menus with HP and EN
+  * Create Attack Menus
+* Built Text Sprite Sheet Functionality
 
 #### v0.0.0-alpha.4
 * Built Battle Outline